Move all DI methods to a single test folder
