+++ AdvancedSystems.Security.Tests/Services/CertificateServiceTests.cs
@@ -1,17 +1,8 @@
 ﻿using System.Linq;
 using System.Security.Cryptography.X509Certificates;
-using System.Threading.Tasks;
 
-using AdvancedSystems.Security.Abstractions;
-using AdvancedSystems.Security.DependencyInjection;
 using AdvancedSystems.Security.Options;
 using AdvancedSystems.Security.Tests.Fixtures;
-
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 
 using Moq;
 
@@ -19,7 +10,7 @@ using Xunit;
 
 namespace AdvancedSystems.Security.Tests.Services;
 
-public class CertificateServiceTests : IClassFixture<CertificateFixture>
+public sealed class CertificateServiceTests : IClassFixture<CertificateFixture>
 {
     private readonly CertificateFixture _sut;
 
@@ -120,72 +111,6 @@ public class CertificateServiceTests : IClassFixture<CertificateFixture>
 
         // Assert
         Assert.Null(certificate);
-    }
-
-    [Fact]
-    public async Task TestAddCertificateService_FromOptions()
-    {
-        // Arrange
-        using var hostBuilder = await new HostBuilder()
-            .ConfigureWebHost(builder => builder
-                .UseTestServer()
-                .ConfigureServices(services =>
-                {
-                    services.AddCertificateService(options =>
-                    {
-                        options.Thumbprint = "123456789";
-                        options.Store = new CertificateStoreOptions
-                        {
-                            Location = StoreLocation.CurrentUser,
-                            Name = StoreName.My,
-                        };
-                    });
-                })
-            .Configure(app =>
-            {
-
-            }))
-            .StartAsync();
-
-        // Act
-        var certificateService = hostBuilder.Services.GetService<ICertificateService>();
-        var certificate = certificateService?.GetConfiguredCertificate();
-
-        // Assert
-        Assert.NotNull(certificateService);
-        Assert.Null(certificate);
-        await hostBuilder.StopAsync();
-    }
-
-    [Fact]
-    public async Task TestAddCertificateService_FromAppSettings()
-    {
-        // Arrange
-        using var hostBuilder = await new HostBuilder()
-            .ConfigureWebHost(builder => builder
-                .UseTestServer()
-                .ConfigureAppConfiguration(config =>
-                {
-                    config.AddJsonFile("appsettings.json", optional: false);
-                })
-                .ConfigureServices((context, services) =>
-                {
-                    services.AddCertificateService(context.Configuration.GetSection(Sections.CERTIFICATE));
-                })
-            .Configure(app =>
-            {
-
-            }))
-            .StartAsync();
-
-        // Act
-        var certificateService = hostBuilder.Services.GetService<ICertificateService>();
-        var certificate = certificateService?.GetConfiguredCertificate();
-
-        // Assert
-        Assert.NotNull(certificateService);
-        Assert.Null(certificate);
-        await hostBuilder.StopAsync();
     }
 
     #endregion
